Review refactor, separte common and visualization utils functions
